+++ src/helpers.ts
@@ -4,21 +4,33 @@ export const buildDesign = (url: string) => ({
 });
 
 type TBindings = {
-  check: string[];
-  prop: Record<string, string | boolean | undefined>;
+  check?: string[];
+  prop?: Record<string, string | boolean | undefined>;
   emit?: string[];
-}
+  custom?: Record<'icon', boolean>;
+};
 
-export const buildSourceBinding = (bindings: TBindings) => (args: any) => {
+const getBinding = (binding: string, map?: string) =>
+  map ? `${map}Prop${binding.charAt(0).toUpperCase() + binding.slice(1)}` : binding;
+
+export const buildSourceBinding = (bindings: TBindings, map?: string) => (args: any) => {
   let result = '';
 
-  Object.keys(bindings.prop).forEach((prop) => {
-    if (args[prop] && args[prop] !== bindings.prop[prop]) result += `${prop}="${args[prop]}" `;
-  });
+  if (bindings.prop) {
+    Object.keys(bindings.prop).forEach((prop) => {
+      if (args[prop] && args[prop] !== bindings.prop![prop]) result += `${prop}="${args[prop]}" `;
+    });
+  }
 
-  bindings.check.forEach((binding) => {
-    if (args[binding]) result += `${binding} `;
-  });
+  if (bindings.custom?.icon) {
+    if (args[getBinding('icon', map)]) result += `:icon="${args[getBinding('icon', map)]}" `;
+  }
+
+  if (bindings.check) {
+    bindings.check.forEach((binding) => {
+      if (args[getBinding(binding, map)]) result += `${binding} `;
+    });
+  }
 
   if (bindings.emit) {
     bindings.emit.forEach((emit) => {
@@ -27,4 +39,4 @@ export const buildSourceBinding = (bindings: TBindings) => (args: any) => {
   }
 
   return result.trim();
-}
+};